Test dtab parsing and pretty-formatting.
Don't bother parsing dtabs before sending them to the server so that the server
deals with proper parsing and sanitation (especially of comments). parseDtab is
now only used in the context of pretty printing so that prefixes are padded so
that each dentry's align. This is not robust (i.e. to accomodate very long
prefixes), but should be sufficient for most server-returned dtabs.